*Complete rewrite* with simplified design of types.
This commit removes lots of code; this is the start of a complete
rewrite. The major change that justifies this is that I did away with
trying to accomplish generic support for many different ublade types.
(Previously, unit blades were representable as unsigned ints, vectors of
ints, or even vectors of symbols). Now, ublades are only represented as
bits in unsigned ints.

(Foreseen) advantages:
* Significantly simplifies the type system, in particular promotion and
  conversion: no need to handle different ublade types generically and
  ensure all operations between all combinations of ublade types work
* Makes type stability far easier to accomplish
* Elimination of much code repetition will improve code readability and
  maintainability

Disadvantages:
* Can no longer represent blades in arbitrarily large dimensions, with
arbitrary basis vectors represented as symbols, as was previously
possible with a ublade type `Vector{Symbol}`. Instead, blades are
limited to `8sizeof(UInt) == 64`. Really, this wasn't that much of an
advantage to begin with --- it was more a curiosity.