Make Blade `bits` a field, not type parameter
The rationale for having bits be a type parameter was originally to keep
open the possibility of dispatch on bits, or perhaps compile-time blade
multiplication.  It ended up being more of a hassle, making functions
which accept `Blade`s recompile for every single combination of bits.

Bits as a field is more in line with the Julian philosophy of what a
type parameter should be -- though the `grade` type parameter is kept.